Implement putIfAbsent and refactor testkit Future waiting methods

+++ src/main/scala/org/infinispan/scala/hotrod/ClientRequest.scala
@@ -2,15 +2,14 @@ package org.infinispan.scala.hotrod
 
 import scala.concurrent.duration.Duration
 
-private[hotrod] sealed abstract class ClientRequest(
-    val code: Byte, val id: Int)
+private[hotrod] sealed abstract class ClientRequest(val code: Int, val id: Int)
 
 private[hotrod] object ClientRequests {
-  case class KeyValue(override val code: Byte, override val id: Int,
+  case class KeyValue(override val code: Int, override val id: Int,
       kv: (Any, Any), lifespan: Duration, maxidle: Duration)
     extends ClientRequest(code, id)
 
-  case class Key(override val code: Byte, override val id: Int, k: Any)
+  case class Key(override val code: Int, override val id: Int, k: Any)
     extends ClientRequest(code, id)
 }
